backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.2
pymongo==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.4
Pillow==10.1.0
tensorflow>=2.16.1
scikit-learn==1.3.2
joblib==1.3.2
pydantic[email]==2.5.0
apscheduler==3.10.4
fastapi-mail==1.4.1